Removing specs for premium caching.
A discussion needs to be had on how premiums should be calculated and
cached.  Premiums were not being cached in a way that was aware of the
difference between coverage start and rate schedule date.  Additionally,
we should be putting a proper fascade around redis access to allow tests
to run even without an active redis server.

For now I have removed the broken code.