Fixing specs.
This spec failure seems to illustrate an invalid cucumber spec.  The
validation on broker_agency_profile.primary_broker_id has been disabled
NOT removed, as this is how it should function.  We need to either
update the spec, write a new one for the correct path, or figure out how
to address the database structure which makes this circular reference
problematic.