Fixed issues with build and emails fields.
Normally you want to exclude the ID fields on new records when doing
nested_attributes_for with mongoid.

If you don't:
- Mongoid assigns an ID as soon as the object is created
- nested_attributes_for, if you assign an object with an ID, will try to
  Mongoid.find that document
- ???
- Crash

What we had previously been doing was constraining the parameters
allowing things into the controller to prevent this.  However, there was
a full whitelist being passed, which allowed the "id" attribute to be
accepted.