this commit fixes plan display based on the variants:
If the shopping household is eligible for -02 or -03, then -02 or -03 variants should display for all metal levels. AI/AN CSR operates different from the other types
If the shopping household is eligible for -04, -05, or -06, this only applies to silver plans. Silver plans should be the CSR variant, all other metal levels should be -01.