Minor changes.
I've moved to checking the database fields for the association so we
don't incur the DB hit of trying to load the association - we just look
at the record we already have. I was nervous about changing the
behaviour of the method in question as well, and wanted to do something
that would leave the logic the same.

I've remove the changes to the template, because those changes create
the possiblity I could have a containing element with no children under
it, which I believe is schema invalid.